Factor the Session class out of the User class.
